# [common] SIMD improvement