--- a/common/arrow/Cargo.toml
+++ b/common/arrow/Cargo.toml
@@ -7,7 +7,6 @@ publish = false
 edition = "2018"
 
 [features]
-default = ["simd"]
 simd = ["arrow/simd"]
 
 [dependencies] # In alphabetical order