--- a/common/datavalues/src/data_array_arithmetic.rs
+++ b/common/datavalues/src/data_array_arithmetic.rs
@@ -11,6 +11,7 @@ use crate::data_array_cast;
 use crate::DataArrayRef;
 use crate::DataColumnarValue;
 use crate::DataType;
+use crate::DataValue;
 use crate::DataValueArithmeticOperator;
 use crate::Float32Array;
 use crate::Float64Array;
@@ -31,25 +32,53 @@ impl DataArrayArithmetic {
         op: DataValueArithmeticOperator,
         left: &DataColumnarValue,
         right: &DataColumnarValue,
-    ) -> Result<DataArrayRef> {
-        let (left_array, right_array) = match (left, right) {
-            (
-                DataColumnarValue::Constant(left_scalar, _),
-                DataColumnarValue::Constant(right_scalar, _),
-            ) => (
-                left_scalar.to_array_with_size(1)?,
-                right_scalar.to_array_with_size(1)?,
-            ),
-            _ => (left.to_array()?, right.to_array()?),
-        };
+    ) -> Result<DataColumnarValue> {
+        match right {
+            DataColumnarValue::Array(right_array) => {
+                let left_array = left.to_array()?;
 
+                Ok(DataColumnarValue::Array(Self::array_array_arithmetic_op(
+                    op,
+                    &left_array,
+                    &right_array,
+                )?))
+            }
+            DataColumnarValue::Constant(right_value, _) => {
+                let mut all_const = false;
+
+                let left_array = match left {
+                    DataColumnarValue::Array(array) => array.clone(),
+                    DataColumnarValue::Constant(scalar, _) => {
+                        all_const = true;
+                        scalar.to_array_with_size(1)?
+                    }
+                };
+                let result = Self::array_scalar_arithmetic_op(op, &left_array, right_value)?;
+
+                if all_const {
+                    let scalar = DataValue::try_from_array(&result, 0)?;
+                    Ok(DataColumnarValue::Constant(scalar, left.len()))
+                } else {
+                    Ok(DataColumnarValue::Array(result))
+                }
+            }
+        }
+    }
+
+    #[inline]
+    fn array_array_arithmetic_op(
+        op: DataValueArithmeticOperator,
+        left_array: &DataArrayRef,
+        right_array: &DataArrayRef,
+    ) -> Result<DataArrayRef> {
         let coercion_type = super::data_type::numerical_arithmetic_coercion(
             &op,
             &left_array.data_type(),
-            &right_array.data_type(),
+            &left_array.data_type(),
         )?;
-        let left_array = data_array_cast(&left_array, &coercion_type)?;
-        let right_array = data_array_cast(&right_array, &coercion_type)?;
+
+        let left_array = data_array_cast(left_array, &coercion_type)?;
+        let right_array = data_array_cast(right_array, &coercion_type)?;
         match op {
             DataValueArithmeticOperator::Plus => {
                 arrow_primitive_array_op!(&left_array, &right_array, &coercion_type, add)
@@ -65,6 +94,34 @@ impl DataArrayArithmetic {
             }
             DataValueArithmeticOperator::Modulo => {
                 arrow_primitive_array_op!(&left_array, &right_array, &coercion_type, modulus)
+            }
+        }
+    }
+
+    #[inline]
+    fn array_scalar_arithmetic_op(
+        op: DataValueArithmeticOperator,
+        left: &DataArrayRef,
+        right_value: &DataValue,
+    ) -> Result<DataArrayRef> {
+        let coercion_type = super::data_type::numerical_arithmetic_coercion(
+            &op,
+            &left.data_type(),
+            &right_value.data_type(),
+        )?;
+        let left_array = data_array_cast(left, &coercion_type)?;
+        let right_value = right_value.cast(&coercion_type)?;
+
+        match op {
+            DataValueArithmeticOperator::Div => {
+                arrow_primitive_array_scalar_op!(left_array, right_value, &coercion_type, divide)
+            }
+            DataValueArithmeticOperator::Modulo => {
+                arrow_primitive_array_scalar_op!(left_array, right_value, &coercion_type, modulus)
+            }
+            _ => {
+                let right_array = right_value.to_array_with_size(left.len())?;
+                Self::array_array_arithmetic_op(op, left, &right_array)
             }
         }
     }