--- a/common/planners/src/plan_expression_chain.rs
+++ b/common/planners/src/plan_expression_chain.rs
@@ -2,84 +2,16 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-use std::fmt;
-
-use common_aggregate_functions::AggregateFunctionFactory;
-use common_aggregate_functions::IAggregateFunction;
-use common_datavalues::DataField;
 use common_datavalues::DataSchemaRef;
-use common_datavalues::DataType;
-use common_datavalues::DataValue;
-use common_exception::ErrorCode;
 use common_exception::Result;
-use common_functions::CastFunction;
 use common_functions::FunctionFactory;
-use common_functions::IFunction;
 
+use crate::ActionAlias;
+use crate::ActionConstant;
+use crate::ActionFunction;
+use crate::ActionInput;
 use crate::Expression;
-
-#[derive(Clone)]
-pub enum ExpressionAction {
-    /// Column which must be in input.
-    Input(ActionInput),
-    /// Constant column with known value.
-    Constant(ActionConstant),
-    Alias(ActionAlias),
-    Function(ActionFunction),
-}
-
-impl fmt::Debug for ExpressionAction {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ExpressionAction::Input(v) => {
-                write!(f, "INPUT {:?}", v.name)
-            }
-            ExpressionAction::Constant(v) => {
-                write!(f, "CONSTANT {:?}", v.name)
-            }
-            ExpressionAction::Alias(v) => {
-                write!(f, "ALIAS {:?}", v.name)
-            }
-            ExpressionAction::Function(v) => {
-                write!(f, "FUNCTION {:?}", v.name)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ActionInput {
-    pub name: String,
-    pub return_type: DataType,
-}
-
-#[derive(Debug, Clone)]
-pub struct ActionConstant {
-    pub name: String,
-    pub value: DataValue,
-}
-
-#[derive(Debug, Clone)]
-pub struct ActionAlias {
-    pub name: String,
-    pub arg_name: String,
-    pub arg_type: DataType,
-}
-
-#[derive(Debug, Clone)]
-pub struct ActionFunction {
-    pub name: String,
-    pub func_name: String,
-    pub return_type: DataType,
-    pub is_aggregated: bool,
-
-    // for functions
-    pub arg_names: Vec<String>,
-    pub arg_types: Vec<DataType>,
-
-    // only for aggregate functions
-    pub arg_fields: Vec<DataField>,
-}
+use crate::ExpressionAction;
 
 #[derive(Debug, Clone)]
 pub struct ExpressionChain {
@@ -245,40 +177,5 @@ impl ExpressionChain {
             }
         }
         Ok(())
-    }
-}
-
-impl ExpressionAction {
-    pub fn column_name(&self) -> &str {
-        match self {
-            ExpressionAction::Input(input) => &input.name,
-            ExpressionAction::Constant(c) => &c.name,
-            ExpressionAction::Alias(a) => &a.name,
-            ExpressionAction::Function(f) => &f.name,
-        }
-    }
-}
-
-impl ActionFunction {
-    pub fn to_function(&self) -> Result<Box<dyn IFunction>> {
-        if self.is_aggregated {
-            return Err(ErrorCode::LogicalError(
-                "Action must be non-aggregated function",
-            ));
-        }
-
-        match self.func_name.as_str() {
-            "cast" => Ok(CastFunction::create(self.return_type.clone())),
-            _ => FunctionFactory::get(&self.func_name),
-        }
-    }
-
-    pub fn to_aggregate_function(&self) -> Result<Box<dyn IAggregateFunction>> {
-        if !self.is_aggregated {
-            return Err(ErrorCode::LogicalError(
-                "Action must be aggregated function",
-            ));
-        }
-        AggregateFunctionFactory::get(&self.func_name, self.arg_fields.clone())
     }
 }