--- a/fusestore/store/Cargo.toml
+++ b/fusestore/store/Cargo.toml
@@ -12,6 +12,7 @@ name = "fuse-store"
 path = "src/bin/fuse-store.rs"
 
 [features]
+default = ["simd"]
 simd = ["common-arrow/simd"]
 
 [dependencies]