I hereby agree to the terms of the CLA available at: https://datafuse.rs/policies/cla/

## Summary
desc: 

```
SELECT sum(number %  3) FROM numbers_mt(10000000000);
before: 1 row in set (13.42 sec)
after: 1 row in set (11.45 sec)



SELECT sum(number /  3) FROM numbers_mt(10000000000);
before: 1 row in set (12.97 sec)
after: 1 row in set (5.71 sec)
```

## Changelog

- Improvement
   - enable SIMD feature by default
   - improve performance for divide and modulus operation when array op scalar value
   - split plan_expression_action and plan_expression_chain. 



## Related Issues

Fixes #788 

## Test Plan

Unit Tests

Stateless Tests