@BohuTANG 2021-06-09
This PR test log is weird:
https://github.com/datafuselabs/datafuse/pull/798/checks?check_run_id=2782733902

The log level is ERROR in the configs, but it shows many INFO logs.
I can not reproduce it in my local machine.